backend/requirements.txt
# ===================================
# FRAMEWORK WEB ET API
# ===================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ===================================
# BASE DE DONNÉES ET ORM
# ===================================
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# ===================================
# SCRAPING
# ===================================
scrapy==2.11.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0

# ===================================
# FILES DE TÂCHES ET CACHE
# ===================================
redis==5.0.1
rq==1.15.1
celery==5.3.4

# ===================================
# NLP ET ANALYSE
# ===================================
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-3.7.0/fr_core_news_md-3.7.0-py3-none-any.whl
nltk==3.8.1
textblob==0.17.1

# ===================================
# ELASTICSEARCH
# ===================================
elasticsearch==8.11.0

# ===================================
# VALIDATION ET SÉRIALISATION
# ===================================
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ===================================
# UTILS ET HELPERS
# ===================================
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2

# ===================================
# LOGS ET MONITORING
# ===================================
loguru==0.7.2
prometheus-client==0.19.0

# ===================================
# TESTS
# ===================================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
factory-boy==3.3.0

# ===================================
# DÉVELOPPEMENT
# ===================================
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0 